refactor: UI components
