+++ src/app/paths.ts
@@ -2,11 +2,8 @@ export const paths = {
   root: '/',
   category: '/category/:id',
   favorites: '/favorites',
-  forgotPassword: '/forgot-password',
   game: '/game',
   login: '/login',
   profile: '/profile',
-  signup: '/signup',
-  updatePassword: '/update-password',
   welcome: '/welcome',
 } as const;
