+++ src/screens/Favorites.tsx
@@ -3,9 +3,7 @@ import { Screen } from 'src/shared/components/Screen';
 export const FavoritesScreen = () => {
   return (
     <Screen>
-      <div className="grid gap-2">
-        <p>Favorites</p>
-      </div>
+      <p>Favorites</p>
     </Screen>
   );
 };
