--- src/screens/LoginScreen.tsx
@@ -1,19 +1,0 @@
-import { Button } from '@nextui-org/react';
-import { loginWithGoogle } from 'src/api/auth/loginWithGoogle';
-import { Screen } from 'src/shared/components/Screen';
-
-export const LoginScreen = () => {
-  return (
-    <Screen>
-      <div className="grid h-full content-center gap-5 p-4">
-        <Button
-          className="justify-self-center"
-          color="primary"
-          onClick={loginWithGoogle}
-        >
-          Login with Google
-        </Button>
-      </div>
-    </Screen>
-  );
-};
