--- src/shared/components/Empty.tsx
@@ -1,9 +1,0 @@
-export const Empty = () => {
-  return (
-    <div className="grid h-20 w-full place-items-center">
-      <div className="grid h-3/4 w-3/4 place-items-center rounded border">
-        <p className="text-center">No data found</p>
-      </div>
-    </div>
-  );
-};
