--- src/shared/components/ErrorMessage.tsx
@@ -1,7 +1,0 @@
-type Props = {
-  message: string;
-};
-
-export const ErrorMessage = ({ message }: Props) => {
-  return <span className="text-red-600">{message}</span>;
-};
