+++ src/shared/components/Screen.tsx
@@ -23,7 +23,11 @@ export const Screen = ({
     <div className="flex h-full flex-col overflow-hidden">
       <main className="flex-1 overflow-auto sm:order-1">
         <div className="mx-auto grid max-w-5xl gap-3 p-5 pt-10">
-          {isLoading ? <Spinner /> : children}
+          {isLoading ? (
+            <Spinner className="justify-self-center p-5" />
+          ) : (
+            children
+          )}
         </div>
       </main>
       {breadcrumbs && (
